Extract string representation of roman units to a read-only list
This way we can easily map the decimal value to the roman string
using the index or `indexOf`, respectively, as long as we account
for the offset for zero-based list indices.

This refactoring will make it easier to add additional numbers to
be parsed, and to implement conversion in both ways (int to string,
string to int) without requiring to implement explicit mapping in
both ways.